+++ host-software/keyplus/layout/ekc_data.py
@@ -30,28 +30,53 @@ class EKCHoldKey(EKCData):
     # Layout option:
     #   1: kc_hold_key
     #   1: delay
-    #   1: options
+    #   1: option_data
     #   1: hold_key
     #   1: tap_key
 
     SIZE = 5 * 2
 
     DEFAULT_DELAY = 200
+    DEFAULT_ACTIVATE_TYPE = 'delay'
 
-    def __init__(self, tap_key=None, hold_key=None, delay=200, options=None, kc_map_function=None):
+    HOLD_KEY_ACTIVATE_DELAY     =  (1 << 0)
+    HOLD_KEY_ACTIVATE_OTHER_KEY =  (1 << 1)
+
+    ACTIVATE_TYPE_MAP = {
+        'delay': (1 << 0),
+        'other_key': (1 << 1),
+    }
+
+
+    def __init__(self, tap_key=None, hold_key=None, delay=200,
+                 activate_type=None, kc_map_function=None):
         self.tap_key = tap_key
         self.hold_key = hold_key
         self.delay = delay
-        self.options = options
         self.kc_map_function = kc_map_function
+        self.activate_type = activate_type or EKCHoldKey.DEFAULT_ACTIVATE_TYPE
 
     def size(self):
         return EKCHoldKey.SIZE
+
+    def check_activate_type_valid(self, type_):
+        if type_ not in EKCHoldKey.ACTIVATE_TYPE_MAP:
+            raise KeyplusSettingsError(
+                "Unknown 'activate_type', got '{}', but expected one of: {}"
+                .format(
+                    self.activate_type,
+                    list(EKCHoldKey.ACTIVATE_TYPE_MAP.keys())
+                )
+            )
 
     def to_bytes(self):
         result = bytearray(EKCHoldKey.SIZE)
 
         option_data = 0
+
+        self.check_activate_type_valid(self.activate_type)
+
+        option_data |= EKCHoldKey.ACTIVATE_TYPE_MAP[self.activate_type]
 
         # _16(KC_HOLD_KEY), \
         # _16(200), \
@@ -106,6 +131,14 @@ class EKCHoldKey(EKCData):
             'delay',
             field_type=int,
             default=EKCHoldKey.DEFAULT_DELAY
+        )
+
+        # Get the delay key field
+        self.activate_type = parser_info.try_get(
+            'activate_type',
+            field_type=str,
+            field_valid_values=EKCHoldKey.ACTIVATE_TYPE_MAP.keys(),
+            default=EKCHoldKey.DEFAULT_ACTIVATE_TYPE,
         )
 
         # Finish parsing `device_name`