+++ src/core/keycode.c
@@ -3,6 +3,10 @@
 
 #include "core/keycode.h"
 
+#include <string.h>
+
+#include "core/error.h"
+
 XRAM flash_ptr_t g_ekc_storage_ptr;
 XRAM uint32_t g_ekc_storage_size;
 
@@ -17,13 +21,15 @@ keycode_t get_ekc_class(keycode_t kc) {
     }
 }
 
-uint8_t get_ekc_data(uint8_t *dest, uint16_t offset, uint16_t size) REENT {
+uint8_t get_ekc_data(void *dest, uint16_t offset, uint16_t size) REENT {
     if (is_valid_storage_pos((flash_ptr_t)g_ekc_storage_ptr + offset) &&
         is_valid_storage_pos((flash_ptr_t)g_ekc_storage_ptr + offset + size - 1)
        ) {
         flash_read(dest, g_ekc_storage_ptr + offset, size);
         return 0;
     } else {
+        register_error(ERROR_EKC_OUT_OF_BOUNDS_ACCESS);
+        memset(dest, 0, size);
         return 1;
     }
 }