+++ src/core/usb_commands.h
@@ -101,9 +101,15 @@ enum {
     RESET_TYPE_SOFTWARE = 1,
 };
 
-#define USB_PRINT_TEXT(text) (usb_print((uint8_t*)text, sizeof(text)))
+#define USB_PRINT_TEXT(text) (usb_print(text, sizeof(text)))
+#define USB_PRINT_VAR(variable) (usb_print(&variable, sizeof(variable)))
+#define USB_PRINT_CONST(value) do { \
+    uint32_t x = value; \
+    usb_print(&x, sizeof(uint32_t)); \
+} while(0)
 
-uint8_t usb_print(const uint8_t *data, uint8_t len);
+uint8_t usb_print(const void* data, uint8_t len);
+
 void reset_usb_reports(void);
 void cmd_send_layer(uint8_t kb_id);
 void handle_vendor_out_reports(void);