+++ src/key_handlers/key_hold.c
@@ -24,6 +24,9 @@
 #define HOLD_KEY_TAP_AUTO_RELEASE_TIME 3
 
 XRAM hold_event_t hold_event_list[MAX_NUM_HOLD_KEYS];
+
+static XRAM bool s_buffer_other_keys;
+
 uint8_t hold_event_list_len;
 
 void handle_hold_keycode(keycode_t keycode, key_event_t event) REENT;
@@ -44,7 +47,7 @@ void hold_key_delete_event(uint8_t i) {
     hold_event_list_len--;
 }
 
-void hold_key_task(void) {
+void hold_key_task(uint8_t other_key_pressed) REENT {
     uint8_t i;
     uint16_t current_time;
 
@@ -57,27 +60,41 @@ void hold_key_task(void) {
     // Check hold key events
     for (i = 0; i < hold_event_list_len; ++i) {
         hold_event_t *hold = &hold_event_list[i];
+        keycode_t keycode;
+        bool timer_passed = has_passed_time16(current_time, hold->end_time);
+
         if (hold->has_generated_event) {
             continue;
         }
 
-        if (has_passed_time16(current_time, hold->end_time)) {
-            keycode_t keycode;
-            if (hold->has_been_tapped) {
-                // this code is used to release tapped keys
-                get_ekc_data(
-                    (uint8_t*)&keycode,
-                    hold->ekc_addr + EKC_OFFSET_TAP_KEYCODE,
-                    sizeof(keycode_t)
-                );
-                queue_keycode_event(keycode, EVENT_RELEASED, hold->kb_id);
-                // need to delete tap event now that it is finished.
-                hold_key_delete_event(i);
-                // since we deleted an event from the list we are processing and
-                // the new event is now at the current list position, need to
-                // decrement here.
-                i--;
-            } else if (!hold->has_been_held) {
+        if (hold->activate_on_other_key && other_key_pressed && !hold->has_been_held) {
+            USB_PRINT_TEXT("hold->pressed, other key");
+
+            // this code is used to press held keys
+            get_ekc_data(
+                (uint8_t*)&keycode,
+                hold->ekc_addr + EKC_OFFSET_HELD_KEYCODE,
+                sizeof(keycode_t)
+            );
+            hold->has_been_held = true;
+            hold->has_generated_event = true;
+            queue_keycode_event(keycode, EVENT_PRESSED, hold->kb_id);
+            s_buffer_other_keys = false;
+        }
+
+        if (other_key_pressed) {
+            USB_PRINT_TEXT("other key done");
+            continue;
+        }
+
+        if (hold->activate_on_delay) {
+            if (!timer_passed) {
+                // if its using the delay activation method, then there's nothing
+                // to do until the delay time has passed
+                continue;
+            }
+
+            if (!hold->has_been_held && !hold->has_been_tapped) {
                 // this code is used to press held keys
                 get_ekc_data(
                     (uint8_t*)&keycode,
@@ -89,7 +106,28 @@ void hold_key_task(void) {
                 queue_keycode_event(keycode, EVENT_PRESSED, hold->kb_id);
             }
         }
+
+        if (hold->has_been_tapped && timer_passed) {
+            USB_PRINT_TEXT("tapped task...");
+            // this code is used to release tapped keys
+            get_ekc_data(
+                &keycode,
+                hold->ekc_addr + EKC_OFFSET_TAP_KEYCODE,
+                sizeof(keycode_t)
+            );
+            queue_keycode_event(keycode, EVENT_RELEASED, hold->kb_id);
+            // need to delete tap event now that it is finished.
+            hold_key_delete_event(i);
+            // since we deleted an event from the list we are processing and
+            // the new event is now at the current list position, need to
+            // decrement here.
+            i--;
+        }
     }
+}
+
+bit_t hold_key_buffer_other_keys(void) {
+    return s_buffer_other_keys;
 }
 
 // The keycode should be the unique address of the key in the external keycode
@@ -103,6 +141,7 @@ void handle_hold_keycode(keycode_t keycode, key_event_t event) REENT {
     if (event == EVENT_RESET) {
         hold_event_list_len = 0;
         hold_keycodes.is_timer_task_active = false;
+        s_buffer_other_keys = false;
         return;
     }
     // } else if (event == EVENT_TIMER_TASK) {
@@ -117,35 +156,54 @@ void handle_hold_keycode(keycode_t keycode, key_event_t event) REENT {
         if (event == EVENT_PRESSED) {
             uint16_t delay;
             uint16_t settings;
+            hold_event_t *hold_key;
 
             if (hold_event_list_len >= MAX_NUM_HOLD_KEYS) {
                 return;
             }
 
-            get_ekc_data((uint8_t*)&delay, this_ekc_addr+EKC_OFFSET_DELAY, sizeof(delay));
-            get_ekc_data((uint8_t*)&settings, this_ekc_addr+EKC_OFFSET_SETTINGS, sizeof(settings));
+            get_ekc_data(&delay, this_ekc_addr+EKC_OFFSET_DELAY, sizeof(delay));
+            get_ekc_data(&settings, this_ekc_addr+EKC_OFFSET_SETTINGS, sizeof(settings));
 
-            hold_event_list[hold_event_list_len].ekc_addr = this_ekc_addr;
-            hold_event_list[hold_event_list_len].end_time = timer_read16_ms() + delay;
-            hold_event_list[hold_event_list_len].kb_id = kb_id;
-            hold_event_list[hold_event_list_len].activate_on_key_press =
-                (uint8_t)settings & HOLD_KEY_SETTING_ACTIVATE_ON_KEY_PRESS;
-            hold_event_list[hold_event_list_len].has_been_held = false;
-            hold_event_list[hold_event_list_len].has_been_tapped = false;
-            hold_event_list[hold_event_list_len].has_generated_event = false;
+            hold_key = &hold_event_list[hold_event_list_len];
+
+            hold_key->ekc_addr = this_ekc_addr;
+            hold_key->end_time = timer_read16_ms() + delay;
+            hold_key->kb_id = kb_id;
+            hold_key->activate_on_delay = (bool)(settings & HOLD_KEY_ACTIVATE_DELAY);
+            hold_key->activate_on_other_key = (bool)(settings & HOLD_KEY_ACTIVATE_OTHER_KEY);
+
+            if (hold_key->activate_on_other_key) {
+                USB_PRINT_TEXT("on_other_key");
+                s_buffer_other_keys = true;
+            }
+            if (hold_key->activate_on_delay) {
+                USB_PRINT_TEXT("on_delay");
+            }
+
+            hold_key->has_been_held = false;
+            hold_key->has_been_tapped = false;
+            hold_key->has_generated_event = false;
 
             hold_event_list_len++;
             hold_keycodes.is_timer_task_active = true;
         } else if (event == EVENT_RELEASED) {
             uint8_t i;
             for (i = 0; i < hold_event_list_len; ++i) {
-                if (hold_event_list[i].ekc_addr == this_ekc_addr &&
-                    hold_event_list[i].kb_id == kb_id ) {
-                    keycode_t keycode;
+                hold_event_t *hold_key = &hold_event_list[i];
+                keycode_t keycode;
+                if (!(hold_key->ekc_addr == this_ekc_addr &&
+                      hold_key->kb_id == kb_id)) {
+                    continue;
+                }
 
-                    if (hold_event_list[i].has_been_held) {
+                if (hold_key->activate_on_delay ||
+                    hold_key->activate_on_other_key
+                ) {
+                    if (hold_key->has_been_held) {
+                        USB_PRINT_TEXT("hold->held_release");
                         get_ekc_data(
-                            (uint8_t*)&keycode,
+                            &keycode,
                             this_ekc_addr + EKC_OFFSET_HELD_KEYCODE,
                             sizeof(keycode_t)
                         );
@@ -155,10 +213,11 @@ void handle_hold_keycode(keycode_t keycode, key_event_t event) REENT {
                         i--; // account for deleted element from this list
                     } else {
                         // tapped key
-                        hold_event_list[i].has_been_tapped = true;
-                        hold_event_list[i].end_time = timer_read16_ms() + HOLD_KEY_TAP_AUTO_RELEASE_TIME;
+                        USB_PRINT_TEXT("hold->tap_release");
+                        hold_key->has_been_tapped = true;
+                        hold_key->end_time = timer_read16_ms() + HOLD_KEY_TAP_AUTO_RELEASE_TIME;
                         get_ekc_data(
-                            (uint8_t*)&keycode,
+                            &keycode,
                             this_ekc_addr + EKC_OFFSET_TAP_KEYCODE,
                             sizeof(keycode_t)
                         );
@@ -167,7 +226,6 @@ void handle_hold_keycode(keycode_t keycode, key_event_t event) REENT {
                         // filled in when the press event was handled.
                         queue_keycode_event(keycode, EVENT_PRESSED, kb_id);
                     }
-
                 }
             }